Rename api-ban helper
