+++ lib/revisions/model.js
@@ -3,7 +3,7 @@ const {omit, pick} = require('lodash')
 const createError = require('http-errors')
 const mongo = require('../util/mongo')
 const Habilitation = require('../habilitations/model')
-const {composeCommune} = require('../util/api-ban-plateforme')
+const {composeCommune} = require('../util/api-ban')
 const {applyValidateBAL} = require('./validate-bal')
 const {notifyPublication} = require('./slack')
